SH.1: unified always-listening demux — route reports by SubscriptionId, fix concurrent round-trip report-loss (#85)
* SH.1: route subscription reports by SubscriptionId (fix latent exchange-keying)

* SH.1: flip actor to a single always-listening demux loop

Replace the conditional-listen run loop + per-request owned recv
(secured_round_trip/secured_read) with one select! that owns recv_from
continuously. Round-trips and reads register a pending oneshot keyed by
(session, exchange); handle_inbound resolves them and routes steady-state
ReportData to subscriptions by SubscriptionId. MRP is centralized in
drive_mrp; Expired resolves the pending op with one transparent
reconnect-and-retry (preserves the prior reconnect-once policy).

Closes the concurrent-round-trip report-loss: a steady report arriving
during an in-flight round-trip is now delivered, not dropped.

matter-commissioning secured_round_trip/secured_read are untouched (still
used by the commissioning driver/examples).

* SH.1: test — concurrent round-trip no longer drops subscription reports

Discriminating guard for M8.5 known limitation #1: a steady-state report
sent while a round-trip is in flight on the same node must be delivered.
Timeout-bounded so a regression fails fast instead of hanging.

* SH.1: CHANGELOG — unified demux + subscription routing fix

* SH.1: reset read/subscribe accumulation on reconnect-retry

on_pending_timeout re-sends the original request on a fresh session; the
retained PendingReply::Read/Subscribe accumulation from the first attempt
must be cleared so the retry starts clean (matches the old fresh-secured_read
retry semantics).

+++ crates/matter-controller/src/subscription.rs
@@ -22,7 +22,7 @@ pub struct AttributeReport {
 pub struct Subscription {
     pub(crate) rx: mpsc::Receiver<AttributeReport>,
     pub(crate) tx: mpsc::Sender<Command>,
-    pub(crate) key: (SessionId, u16),
+    pub(crate) key: (SessionId, u32),
     pub(crate) cancelled: bool,
 }
